Move common functions to a helper module
